Parse Version 1: Currently, parses and store Athlete data
Athlete class, as well as TrackProgram name have been created, as well as a test Environment for testing things. More importantly the track Program is not optimized at all, need to create a small temporary object and or boolean value with a switch statement for when it comes to checking.