Created GUI and results are copied onto Clipboard
GUI for taking input of URL and School name with a checkbox asking if a name contains a comma has been created. The result is immediately copied onto the clipboard and structured in a way where pasting into an Excel sheet or Google sheet automatically puts it into rows and columns correctly.