Version 1.1.0 - Added TextArea to see submission, functionally to go back
Added many features such as textarea to see what exactly was copied, added a dropdown menu that shows priors submissions, also cleaned up some of the repetition and organized the TrackProgram code a bit more, also added some validation using patterns and regex to ensure the address is correct. Also added a clear button that clears the inputs but not the submissions.