All Test Cases Finished(for now), Collected Important Groups
Group 4 Checks if there is a comma.
Group 6 Tends to be the third name(need more testing).
Group 15 Tends to be either null or the event, usually used when there are
two times such as a "seed" and the actual time
IE      Seeded 2:25 , Actual 2:17.
Group 12 Tends to either the event or the seed, usually used normally however some meets have seed times which is where group 15 comes into play.
Plan to insert into the TrackProgram.java and restructure that.
Also need to remove comma option from the GUI considering the application can now do it on it's own.