Currently creates and writes into a directory, creates csv file
CSV file is correctly formatted and is created, however, it is currently stuck to submission 0, this has to do with the fact that it is operating on separate self creating threads, an easy fix is to simply just have a static value carry the submission numbers.
Also need to properly set up directory checks to see if the name already exists.
Also need to check if the user wants to create their own directory somewhere else, this would require some gui as well.