Set up the PermissionMenu, as well created ForcedException
PermissionMenu was created and while it is not completely done, it is currently functional however, the code itself could be made more efficient.
Created the Forced Exception which will allow the user to have a pop up window where they enter information about why they are reporting a problem which will also showcase how to submit the problem.
Would have finished fixing the PermissionMenu and then commited but need to work on the laptop so committing now would be easier.